+++ cutils/src/main/java/com/carlos/cutils/aop/CSingleClickAspect.kt
@@ -17,7 +17,6 @@ class CSingleClickAspect {
 
     /**
      * 定义切点，标记切点为所有被@SingleClick注解的方法
-     * 注意：这里me.baron.test.annotation.SingleClick需要替换成你自己项目中SingleClick这个类的全路径哦
      */
     @Pointcut("execution(@com.carlos.cutils.aop.CSingleClick * *(..))")
     fun methodAnnotated() {}