+++ cutils/src/main/java/com/carlos/cutils/base/fragment/CBaseFragment.kt
@@ -21,8 +21,12 @@ abstract class CBaseFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         val view = inflater.inflate(layoutId(), container, false)
-        initView(view)
         return view
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initView(view)
     }
 
 }