+++ demo/src/main/java/com/carlos/cutils/demo/fragment/WebviewFragment.kt
@@ -1,0 +1,8 @@
+package com.carlos.cutils.demo.fragment
+
+import com.carlos.cutils.base.fragment.CBaseWebFragment
+
+/**
+ * Created by Carlos on 2019/3/20.
+ */
+class WebviewFragment : CBaseWebFragment("https://www.jianshu.com/p/e1099a94b979")